F01: Add implementation of DepartementService
