+++ src/main/java/com/carto/sn/service/DepartementServiceImpl.java
@@ -1,5 +1,25 @@
 package com.carto.sn.service;
 
-public class DepartementServiceImpl {
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
+import com.carto.sn.dao.DepartementRepository;
+import com.carto.sn.entities.Departement;
+
+public class DepartementServiceImpl implements DepartementService {
+	
+	@Autowired
+	private DepartementRepository departementRepository;
+	
+	@Override
+	public List<Departement> tousLesDepartements() {
+		return departementRepository.findAll();
+	}
+
+	@Override
+	public List<Departement> findDepartementByNomRegion(String nomRegion) {
+		return departementRepository.findDepartementByNomRegion(nomRegion);
+	}
 
 }
